faster local node and udpate delegate test
